+++ pdf/lib/widgets/document.dart
@@ -16,21 +16,26 @@
 
 part of widget;
 
-abstract class BasePage {
-  BasePage({@required this.pageFormat}) : assert(pageFormat != null);
-
-  final PdfPageFormat pageFormat;
-
-  @protected
-  void generate(Document document);
-}
-
 class Document {
   Document(
       {PdfPageMode pageMode = PdfPageMode.none,
       DeflateCallback deflate,
-      this.theme})
-      : document = PdfDocument(pageMode: pageMode, deflate: deflate);
+      this.theme,
+      String title,
+      String author,
+      String creator,
+      String subject,
+      String keywords,
+      String producer})
+      : document = PdfDocument(pageMode: pageMode, deflate: deflate) {
+    document.info = PdfInfo(document,
+        title: title,
+        author: author,
+        creator: creator,
+        subject: subject,
+        keywords: keywords,
+        producer: producer);
+  }
 
   static bool debug = false;
 
@@ -38,7 +43,7 @@ class Document {
 
   final Theme theme;
 
-  void addPage(BasePage page) {
+  void addPage(Page page) {
     page.generate(this);
   }
 
@@ -48,50 +53,92 @@ class Document {
 typedef BuildCallback = Widget Function(Context context);
 typedef BuildListCallback = List<Widget> Function(Context context);
 
-class Page extends BasePage {
-  Page(
-      {PdfPageFormat pageFormat = PdfPageFormat.a4,
+enum PageOrientation { natural, landscape, portrait }
+
+class Page {
+  const Page(
+      {this.pageFormat = PdfPageFormat.standard,
       BuildCallback build,
       this.theme,
+      this.orientation = PageOrientation.natural,
       EdgeInsets margin})
-      : margin = margin ??
-            EdgeInsets.fromLTRB(pageFormat.marginLeft, pageFormat.marginTop,
-                pageFormat.marginRight, pageFormat.marginBottom),
-        _build = build,
-        super(pageFormat: pageFormat);
+      : assert(pageFormat != null),
+        _margin = margin,
+        _build = build;
 
-  final EdgeInsets margin;
+  final PdfPageFormat pageFormat;
+
+  final PageOrientation orientation;
+
+  final EdgeInsets _margin;
 
   final BuildCallback _build;
 
   final Theme theme;
 
+  bool get mustRotate =>
+      (orientation == PageOrientation.landscape &&
+          pageFormat.height > pageFormat.width) ||
+      (orientation == PageOrientation.portrait &&
+          pageFormat.width > pageFormat.height);
+
+  EdgeInsets get margin {
+    if (_margin != null) {
+      if (mustRotate) {
+        return EdgeInsets.fromLTRB(
+            _margin.bottom, _margin.left, _margin.top, _margin.right);
+      } else {
+        return _margin;
+      }
+    }
+
+    if (mustRotate) {
+      return EdgeInsets.fromLTRB(pageFormat.marginBottom, pageFormat.marginLeft,
+          pageFormat.marginTop, pageFormat.marginRight);
+    } else {
+      return EdgeInsets.fromLTRB(pageFormat.marginLeft, pageFormat.marginTop,
+          pageFormat.marginRight, pageFormat.marginBottom);
+    }
+  }
+
+  @protected
   void debugPaint(Context context) {
+    final EdgeInsets _margin = margin;
     context.canvas
-      ..setFillColor(PdfColor.lightGreen)
+      ..setFillColor(PdfColors.lightGreen)
       ..moveTo(0, 0)
       ..lineTo(pageFormat.width, 0)
       ..lineTo(pageFormat.width, pageFormat.height)
       ..lineTo(0, pageFormat.height)
-      ..moveTo(margin.left, margin.bottom)
-      ..lineTo(margin.left, pageFormat.height - margin.top)
-      ..lineTo(pageFormat.width - margin.right, pageFormat.height - margin.top)
-      ..lineTo(pageFormat.width - margin.right, margin.bottom)
+      ..moveTo(_margin.left, _margin.bottom)
+      ..lineTo(_margin.left, pageFormat.height - _margin.top)
+      ..lineTo(
+          pageFormat.width - _margin.right, pageFormat.height - _margin.top)
+      ..lineTo(pageFormat.width - _margin.right, _margin.bottom)
       ..fillPath();
   }
 
-  @override
+  @protected
   void generate(Document document) {
     final PdfPage pdfPage = PdfPage(document.document, pageFormat: pageFormat);
     final PdfGraphics canvas = pdfPage.getGraphics();
-    final BoxConstraints constraints = BoxConstraints(
-        maxWidth: pageFormat.width, maxHeight: pageFormat.height);
+    final EdgeInsets _margin = margin;
+    final BoxConstraints constraints = mustRotate
+        ? BoxConstraints(
+            maxWidth: pageFormat.height - _margin.vertical,
+            maxHeight: pageFormat.width - _margin.horizontal)
+        : BoxConstraints(
+            maxWidth: pageFormat.width - _margin.horizontal,
+            maxHeight: pageFormat.height - _margin.vertical);
 
     final Theme calculatedTheme = theme ?? document.theme ?? Theme.base();
     final Map<Type, Inherited> inherited = <Type, Inherited>{};
     inherited[calculatedTheme.runtimeType] = calculatedTheme;
-    final Context context =
-        Context(page: pdfPage, canvas: canvas, inherited: inherited);
+    final Context context = Context(
+        document: document.document,
+        page: pdfPage,
+        canvas: canvas,
+        inherited: inherited);
     if (_build != null) {
       final Widget child = _build(context);
       layout(child, context, constraints);
@@ -103,19 +150,12 @@ class Page extends BasePage {
   void layout(Widget child, Context context, BoxConstraints constraints,
       {bool parentUsesSize = false}) {
     if (child != null) {
-      final BoxConstraints childConstraints = BoxConstraints(
-          minWidth: constraints.minWidth,
-          minHeight: constraints.minHeight,
-          maxWidth: constraints.hasBoundedWidth
-              ? constraints.maxWidth - margin.horizontal
-              : margin.horizontal,
-          maxHeight: constraints.hasBoundedHeight
-              ? constraints.maxHeight - margin.vertical
-              : margin.vertical);
-      child.layout(context, childConstraints, parentUsesSize: parentUsesSize);
+      final EdgeInsets _margin = margin;
+      child.layout(context, constraints, parentUsesSize: parentUsesSize);
+      assert(child.box != null);
       child.box = PdfRect(
-          margin.left,
-          pageFormat.height - child.box.height - margin.top,
+          _margin.left,
+          pageFormat.height - child.box.height - _margin.top,
           child.box.width,
           child.box.height);
     }
@@ -130,140 +170,24 @@ class Page extends BasePage {
       return true;
     }());
 
-    if (child != null) {
-      child.paint(context);
-    }
-  }
-}
-
-class MultiPage extends Page {
-  MultiPage(
-      {PdfPageFormat pageFormat = PdfPageFormat.a4,
-      BuildListCallback build,
-      this.crossAxisAlignment = CrossAxisAlignment.start,
-      this.header,
-      this.footer,
-      Theme theme,
-      EdgeInsets margin})
-      : _buildList = build,
-        super(pageFormat: pageFormat, margin: margin, theme: theme);
-
-  final BuildListCallback _buildList;
-
-  final CrossAxisAlignment crossAxisAlignment;
-
-  final BuildCallback header;
-
-  final BuildCallback footer;
-
-  @override
-  void generate(Document document) {
-    if (_buildList == null) {
+    if (child == null) {
       return;
     }
 
-    final BoxConstraints constraints = BoxConstraints(
-        maxWidth: pageFormat.width, maxHeight: pageFormat.height);
-    final BoxConstraints childConstraints =
-        BoxConstraints(maxWidth: constraints.maxWidth - margin.horizontal);
-    final Theme calculatedTheme = theme ?? document.theme ?? Theme.base();
-    final Map<Type, Inherited> inherited = <Type, Inherited>{};
-    inherited[calculatedTheme.runtimeType] = calculatedTheme;
-    Context context;
-    double offsetEnd;
-    double offsetStart;
-    int index = 0;
-    final List<Widget> children = _buildList(Context(inherited: inherited));
-    WidgetContext widgetContext;
-
-    while (index < children.length) {
-      final Widget child = children[index];
-
-      if (context == null) {
-        final PdfPage pdfPage =
-            PdfPage(document.document, pageFormat: pageFormat);
-        final PdfGraphics canvas = pdfPage.getGraphics();
-        context = Context(page: pdfPage, canvas: canvas, inherited: inherited);
-        assert(() {
-          if (Document.debug) {
-            debugPaint(context);
-          }
-          return true;
-        }());
-        offsetStart = pageFormat.height - margin.top;
-        offsetEnd = margin.bottom;
-        if (header != null) {
-          final Widget headerWidget = header(context);
-          if (headerWidget != null) {
-            headerWidget.layout(context, childConstraints,
-                parentUsesSize: false);
-            headerWidget.box = PdfRect(
-                margin.left,
-                offsetStart - headerWidget.box.height,
-                headerWidget.box.width,
-                headerWidget.box.height);
-            headerWidget.paint(context);
-            offsetStart -= headerWidget.box.height;
-          }
-        }
-
-        if (footer != null) {
-          final Widget footerWidget = footer(context);
-          if (footerWidget != null) {
-            footerWidget.layout(context, childConstraints,
-                parentUsesSize: false);
-            footerWidget.box = PdfRect(margin.left, margin.bottom,
-                footerWidget.box.width, footerWidget.box.height);
-            footerWidget.paint(context);
-            offsetEnd += footerWidget.box.height;
-          }
-        }
-      }
-
-      if (widgetContext != null && child is SpanningWidget) {
-        child.restoreContext(widgetContext);
-        widgetContext = null;
-      }
-
-      child.layout(context, childConstraints, parentUsesSize: false);
-
-      if (offsetStart - child.box.height < offsetEnd) {
-        if (child.box.height <= pageFormat.height - margin.vertical) {
-          context = null;
-          continue;
-        }
-
-        if (!(child is SpanningWidget)) {
-          throw Exception(
-              'Widget won\'t fit into the page as its height (${child.box.height}) '
-              'exceed a page height (${pageFormat.height - margin.vertical}). '
-              'You probably need a SpanningWidget or use a single page layout');
-        }
-
-        final SpanningWidget span = child;
-
-        child.layout(context,
-            childConstraints.copyWith(maxHeight: offsetStart - offsetEnd),
-            parentUsesSize: false);
-        child.box = PdfRect(margin.left, offsetStart - child.box.height,
-            child.box.width, child.box.height);
-        child.paint(context);
-
-        if (span.canSpan) {
-          widgetContext = span.saveContext();
-        } else {
-          index++;
-        }
-
-        context = null;
-        continue;
-      }
-
-      child.box = PdfRect(margin.left, offsetStart - child.box.height,
-          child.box.width, child.box.height);
+    if (mustRotate) {
+      final EdgeInsets _margin = margin;
+      final Matrix4 mat = Matrix4.identity();
+      mat
+        ..rotateZ(-math.pi / 2)
+        ..translate(-pageFormat.height - _margin.left + _margin.top,
+            child.box.height - child.box.width + _margin.left - _margin.bottom);
+      context.canvas
+        ..saveContext()
+        ..setTransform(mat);
       child.paint(context);
-      offsetStart -= child.box.height;
-      index++;
+      context.canvas.restoreContext();
+    } else {
+      child.paint(context);
     }
   }
 }