+++ pdf/lib/widgets/widget.dart
@@ -18,7 +18,13 @@ part of widget;
 
 @immutable
 class Context {
-  const Context({this.page, this.canvas, this.inherited});
+  const Context({
+    @required this.document,
+    this.page,
+    this.canvas,
+    @required this.inherited,
+  })  : assert(document != null),
+        assert(inherited != null);
 
   final PdfPage page;
 
@@ -26,11 +32,19 @@ class Context {
 
   final Map<Type, Inherited> inherited;
 
-  int get pageNumber => page.pdfDocument.pdfPageList.pages.indexOf(page) + 1;
+  final PdfDocument document;
+
+  int get pageNumber => document.pdfPageList.pages.indexOf(page) + 1;
+
+  int get pagesCount => document.pdfPageList.pages.length;
 
   Context copyWith(
-      {PdfPage page, PdfGraphics canvas, Map<Type, Inherited> inherited}) {
+      {PdfPage page,
+      PdfGraphics canvas,
+      Matrix4 ctm,
+      Map<Type, Inherited> inherited}) {
     return Context(
+        document: document,
         page: page ?? this.page,
         canvas: canvas ?? this.canvas,
         inherited: inherited ?? this.inherited);
@@ -48,13 +62,19 @@ class Inherited {}
 abstract class Widget {
   Widget();
 
+  /// The bounding box of this widget, calculated at layout time
   PdfRect box;
 
+  /// First widget pass to calculate the children layout and
+  /// bounding [box]
   @protected
   void layout(Context context, BoxConstraints constraints,
       {bool parentUsesSize = false});
 
+  /// Draw itself and its children, according to the calculated
+  /// [box.offset]
   @protected
+  @mustCallSuper
   void paint(Context context) {
     assert(() {
       if (Document.debug) {
@@ -67,22 +87,10 @@ abstract class Widget {
   @protected
   void debugPaint(Context context) {
     context.canvas
-      ..setStrokeColor(PdfColor.purple)
+      ..setStrokeColor(PdfColors.purple)
       ..drawRect(box.x, box.y, box.width, box.height)
       ..strokePath();
   }
-}
-
-class WidgetContext {}
-
-abstract class SpanningWidget extends Widget {
-  bool get canSpan => false;
-
-  @protected
-  WidgetContext saveContext();
-
-  @protected
-  void restoreContext(WidgetContext context);
 }
 
 abstract class StatelessWidget extends Widget {
@@ -97,6 +105,7 @@ abstract class StatelessWidget extends Widget {
 
     if (_child != null) {
       _child.layout(context, constraints, parentUsesSize: parentUsesSize);
+      assert(_child.box != null);
       box = _child.box;
     } else {
       box = PdfRect.fromPoints(PdfPoint.zero, constraints.smallest);
@@ -132,16 +141,15 @@ abstract class SingleChildWidget extends Widget {
       {bool parentUsesSize = false}) {
     if (child != null) {
       child.layout(context, constraints, parentUsesSize: parentUsesSize);
+      assert(child.box != null);
       box = child.box;
     } else {
       box = PdfRect.fromPoints(PdfPoint.zero, constraints.smallest);
     }
   }
 
-  @override
-  void paint(Context context) {
-    super.paint(context);
-
+  @protected
+  void paintChild(Context context) {
     if (child != null) {
       final Matrix4 mat = Matrix4.identity();
       mat.translate(box.x, box.y);